refactor(cli): flatten cli/repl into cli package
Drop the cli/repl subpackage: REPL, NewREPL, the helper formatters, and
the load/save dot commands now live directly under cli. The dedicated
repl.Option / repl.WithFS pair, which existed solely to avoid colliding
with cli.Option / cli.WithFS across the package boundary, is replaced
by NewREPL(in, out, fs WriteFS) — a single direct argument since fs was
the only option. Pass nil to disable .load and .save when callers
construct the REPL standalone; cli.Root keeps wiring cli.OS() in for
the default minivm command.
